+++ modules/newton_method/test/test_newton_method.cpp
@@ -5,7 +5,7 @@
 #include "include/Newton_method.h"
 
 const double eps = 0.01;
-const size_t polynom_size = 10;
+const unsigned int polynom_size = 10;
 
 TEST(test_newton_method, can_generate_random_polynom) {
     // Arrange
@@ -18,7 +18,7 @@ TEST(test_newton_method, can_generate_random_polynom) {
 
 TEST(test_newton_method, random_polynom_whith_zero_size) {
     // Arrange
-    size_t zero_size = 0;
+    unsigned int zero_size = 0;
     double *polynom_coef = new double[zero_size];
     NewtonMethod nm;
     // Act & Assert
@@ -54,7 +54,7 @@ TEST(test_newton_method, work_with_random_polynom) {
 TEST(test_newton_method, work_with_random_polynom_zero_size) {
     // Arrange
     NewtonMethod nm;
-    size_t zero_size = 0;
+    unsigned int zero_size = 0;
     double *polynom_coef = new double[polynom_size];
     nm.GeneateRandomPolynom(zero_size, polynom_coef);
     // Act & Assert
@@ -83,10 +83,11 @@ TEST(test_newton_method, check_result_function) {
 TEST(test_newton_method, check_result_polynom) {
     // Arrange
     double *polynom_coef = new double[polynom_size];
-    NewtonMethod nm;	
+    NewtonMethod nm;
     nm.GeneateRandomPolynom(polynom_size, polynom_coef);
     double root = nm.NewtonMethodPolynom(eps, polynom_size, polynom_coef);
     // Act & Assert
-    ASSERT_TRUE(nm.CheckRootPolynom(root, eps*100, polynom_size, polynom_coef) );
+    ASSERT_TRUE(nm.CheckRootPolynom(root, eps*100, polynom_size,
+                                                               polynom_coef));
     delete[]polynom_coef;
 }
